+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Allow the foodsaver, which is verifiable, to create itself in the profile settings (without cut markers) !2198 @chriswalg
 - Dashboard revamped, shines in a new style and new features !2227 !2245 !2248  @sefn
 - In the footer of the website, "For developers" is added under "Participate" to create an easier way to IT. !2241 @NINI1988
+- On the login page, "Stay logged in for one day" is saved in local storage for the next logins. !2249 @NINI1988
 
 ## Changes
 -Checks in the id card creation function if only one foodsaver is selected and then reduces the